+++ test/Agent.test.js
@@ -195,4 +195,6 @@ describe('Agent', function() {
     });
   });
 
+  // TODO: test extendTo
+
 });