+++ HISTORY.md
@@ -10,7 +10,7 @@
 - Implemented support for connecting with an alternative id.
 - Implemented support for sending a message via a specific transport, by
   specifying either the transports type or id.
-- Extended Agent with a function `extend`, for easy loading of modules. 
+- Extended Agent with a functions `extend`, and `extentTo` for module loading. 
 - Moved pattern listening functionality in a separate module.
 - Renamed `Actor` to `Agent`, and `peer` to `agent`. 
 - Renamed `*MessageBus` to `*Transport`.